requirements.txt
django==4.0.6
python-dotenv==0.20.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
psycopg2-binary==2.9.3
djoser==2.1.0
channels==3.0.5